# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Survivor League API
  description: >
    REST API for the College Football Survivor Pool app. This spec is the
    source of truth for the shared TypeScript client generated into
    packages/api-client (consumed by both web and mobile).

    Phase 1 added auth (register/login/refresh/logout) and the current-user
    endpoints. Phase 2 added league CRUD, membership, the invite-code join
    flow, and the conferences list. Phase 3 added read-only schedule
    endpoints (weeks/games/teams, populated by a CFBD sync) and the first
    site-admin endpoints (triggering/viewing that sync). Phase 4 added pick
    submission/retrieval with server-enforced per-game locking and the
    no-repeat-team rule, plus pre-lock pick-hiding for other league members.
    Phase 5 adds the leaderboard endpoint — the grading/elimination pipeline
    and the live poll loop that drives it are internal (internal/grading,
    internal/livepoll), not client-facing API surface. Phase 6 adds the
    commissioner buy-back endpoint (one-time-ever reinstatement of an
    eliminated member). Phase 7 adds notifications: device-token
    registration and notification preferences (the notification_outbox
    dispatcher, Expo Push/Resend delivery, and the five trigger types —
    pick_reminder, eliminated, survived, mass_wipeout, buyback — are
    internal, internal/notify, not client-facing API surface beyond these
...
    (GET /admin/leagues, GET /admin/users), user disable/enable, a
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /leagues/{id}/weeks/{weekId}/available-teams:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: weekId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: listAvailableTeams
      summary: List pickable teams for a week
      description: >
        Requires a non-removed membership in the league (requireLeagueMember).
        Every team in the league's locked conference that has a game in the
        given week, each annotated with is_locked (that specific game has
        kicked off) and is_used_elsewhere (this team is already the
        requester's pick for a DIFFERENT week, regardless of that other
        pick's lock status). Also surfaces the requester's current pick for
        THIS week (if any) so the UI can render the current selection
        without a second round-trip.
      tags:
        - Picks
      responses:
        "200":
          description: This week's pickable teams, plus the requester's current pick for it
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailableTeamsResponse"
        "403":
          description: Not a member of this league
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: League or week not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /leagues/{id}/weeks/{weekId}/picks:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: weekId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: listWeekPicks
      summary: List every league member's pick status for a week
      description: >
        Requires a non-removed membership in the league (requireLeagueMember).
        Privacy rule: for every membership that is NOT the requester's own,
        game_id/team_id are included only once that pick's game has kicked
        off — omitted as a pair (never just team_id) while it hasn't, since
        revealing even which game a member is tied to narrows their choice
        to two teams. The requester's own entry always includes full detail
        when has_picked is true, regardless of lock status.
      tags:
        - Picks
      responses:
        "200":
          description: Every non-removed member's pick status for this week
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberPickStatus"
        "403":
          description: Not a member of this league
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: League or week not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /teams:
    get:
      operationId: listTeams
      summary: List FBS teams
      description: >
        Populated by the CFBD schedule sync. `conference` filters to an
        exact (case-sensitive) match against the canonical FBS conference
        names (see GET /conferences); omit it to list every FBS team across
        every conference.
      tags:
        - Schedule
      parameters:
        - name: conference
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Teams matching the filter, ordered by name
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
        "401":
          description: Missing, invalid, or expired access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
    Team:
      type: object
      required:
        - id
        - external_id
        - name
        - conference
      properties:
        id:
          type: string
          format: uuid
        external_id:
          type: string
          description: CFBD's team id, stored as-is (no id-mangling).
        name:
          type: string
        conference:
          type: string
          description: >
            The *normalized* conference name (see
            internal/schedule/conferences.go) — always one of the canonical
            names returned by GET /conferences, never CFBD's raw string.
        logo_url:
          type: string
...
    Pick:
      type: object
      description: A single week's pick. `locked` is computed live against the backing game's kickoff_at, not stored.
      required:
        - game_id
        - team_id
        - locked
      properties:
        game_id:
          type: string
          format: uuid
        team_id:
          type: string
          format: uuid
        locked:
          type: boolean
...
    AvailableTeam:
      type: object
      description: One pickable team for a week, with its opponent/game inline so the picks screen doesn't need N+1 lookups.
      required:
        - team_id
        - team_name
        - opponent_team_id
        - opponent_name
        - game_id
        - kickoff_at
        - is_home
        - is_locked
        - is_used_elsewhere
        - is_current_pick
      properties:
        team_id:
          type: string
          format: uuid
        team_name:
          type: string
        team_logo_url:
          type: string
        opponent_team_id:
          type: string
          format: uuid
        opponent_name:
          type: string
        opponent_logo_url:
          type: string
        game_id:
          type: string
          format: uuid
        kickoff_at:
          type: string
          format: date-time
        is_home:
          type: boolean
          description: True if this team is the home team in this game.
        is_locked:
          type: boolean
          description: This team's game has already kicked off.
        is_used_elsewhere:
          type: boolean
          description: This team is already the requester's pick for a DIFFERENT week (regardless of that pick's lock status).
        is_current_pick:
          type: boolean
          description: This team is the requester's current pick for THIS week.
    AvailableTeamsResponse:
      type: object
      required:
        - teams
      properties:
        teams:
          type: array
          items:
            $ref: "#/components/schemas/AvailableTeam"
        current_pick:
          $ref: "#/components/schemas/Pick"
          description: The requester's current pick for this week, if any (absent if they haven't picked yet).
    MemberPickStatus:
      type: object
      description: >
        A league member's pick status for a week. game_id/team_id are only
        present when the caller is entitled to see them — see
        GET /leagues/{id}/weeks/{weekId}/picks's privacy-rule description.
      required:
        - membership_id
        - display_name
        - has_picked
      properties:
        membership_id:
          type: string
          format: uuid
        display_name:
          type: string
        has_picked:
          type: boolean
        game_id:
          type: string
          format: uuid
        team_id:
          type: string
          format: uuid
...